pyproject.toml
[tool.poetry]
name = "potnia"
version = "0.3.0"
description = "Potnia is an open-source Python library designed to convert Romanized transliterations of ancient texts into Unicode representations of ther respective native scripts.."
authors = ["Emily Tour <esmtour@gmail.com>", "Kabir Manandhar Shrestha", "Robert Turnbull"]
license = "Apache-2.0"
readme = "README.rst"
repository = "https://github.com/AncientNLP/potnia/"
documentation = "https://AncientNLP.github.io/potnia"
homepage = "https://github.com/AncientNLP/potnia/"
keywords = ["linear a", "linear b", "sumerian", "akkadian", "cuneiform", "unicode"]
# For classifiers see https://pypi.org/classifiers/
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Science/Research",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
typer = "^0.12.5"
guigaga = ">=0.0.5"
numpy = "<2"
pybtex = ">=0.24.0"
pybtexnbib = ">=0.1.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
coverage = ">=7.4.3"
Sphinx = ">=5.0.0"
sphinx-rtd-theme = ">=1.0.0"
sphinx-autobuild = ">=2021.3.14"
sphinx-copybutton = ">=0.4.0"
myst-parser = "^3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
potnia = "potnia.main:app"
